Convert `addCardToDashboard` to TypeScript
